Update config.yaml
Revised the copyright statement, however for now displaying a creative commons license as it is all my own work. Needed to revise the outputs otherwise the site would not render (tried clearing the cache and re-adding and still didn't work). This may be an issue so noting it in the changelog.