Update menus.yaml
Simplified the site further by removing talks and blog link (have added blog link to footer instead as it is not immediately relevant to this portfolio). This is a minimum viable product and I will add a resume download for this stuff until I have built it well and accessibly.