Update about.html
Changed the order of content so that gravitar appears below the name and title. Added accurate alt text to my gravitar image. Changed the social-media icons to include an aria-label and a title as they were aria-hidden but were focusable. Didn't think that this was quite right by may need further tweaking. Added a custom class for the h2 about so that the styling was consistent with the rest of the page. Made sure that the section sub-headings for interests and education were actual headings (h3).